refactor(06): extract function loadComp to module
