+++ 06-build-page/builder/plugins/html-plugin.js
@@ -1,23 +1,6 @@
 const path = require('path');
 const Loader = require('../utilities/loader');
-
-/**
- * @param {String} sourcePath 
- * @returns {Promise}
- */
-async function loadComp(sourcePath) {
-  const components = {};
-  let names = [];
-  await new Loader(sourcePath).readDir((data) => names=data).untilResolve();
-
-  await Promise.all(names.map((name) => {
-    let data = '';
-    return new Loader(path.join(sourcePath, name)).readFile((chunk)=>data+=chunk).then(()=>{
-      components[path.parse(name).name] = data;
-    }).untilResolve()
-  }));
-  return components;
-}
+const loadComp = require('../utilities/load-comp')
 
 module.exports = class HtmlPlugin {
   path = {};
